\name{nvis.keyword}
\alias{nvis.keyword}
\title{nvis keyword}
\usage{nvis.keyword(key)}
\description{Keyword function to match "string" description to proper object of analysis}
\author{Fernando Fuentes \email{fuentes.f4@gmail.com}}
\keyword{methods}
\arguments{
\item{key}{string keyword to be deciphered}
}
\examples{\dontrun{
# example of use
nvis.keyword("inm")
# returns:
#[1] "RHO" "W"   "K"   "M0i" "A"   "L"   "DK"  "M1i"
}}
