\name{Syrme dataset with 189 forces}
\docType{data}
\alias{skyrme_189}
\title{Skyrme surface/volume parameters and infinite nuclear matter properties}
\description{
A dataset containing the surface parameters, volume parameters and 
infinite nuclear matter properties for 189 forces of skyrme energy density functional.
}
\usage{data(skyrme_189)}
\format{A data frame with 189 rows and 27 variables}
\keyword{datasets}
