\name{Syrme dataset with 65 forces}
\docType{data}
\alias{skyrme_65}
\title{Skyrme surface/volume parameters and infinite nuclear matter properties}
\description{
A dataset containing the surface parameters, volume parameters and 
infinite nuclear matter properties for 65 forces of skyrme energy density functional.
}
\usage{data(skyrme_65)}
\format{A data frame with 65 rows and 27 variables}
\keyword{datasets}
